edit permission (guest,member,admin)

+++ app/Http/Controllers/CartController.php
@@ -15,8 +15,15 @@ class CartController extends Controller
      */
     public function index()
     {
+        // mengambil data user yang sedang login
         $user = Auth::user();
-        // dd($user);
+        // mengecek data user, karena hanya member/admin yang boleh berkunjung
+        if($user==null)
+            return view('auth.login');
+        // mengecek data user, karena hanya member yang boleh menambah data cart
+        if($user->user!=2)
+            abort(403);
+        // mengirim data user ke halaman cart
         return view('pizza.cart', compact('user'));
     }
 
@@ -38,8 +45,18 @@ class CartController extends Controller
      */
     public function store(Request $request, $id)
     {
+        // mengambil data user yang sedang login
         $user = Auth::user();
+        // mengecek data user, karena hanya member/admin yang boleh berkunjung
+        if($user==null)
+            return view('auth.login');
+        // mengecek data user, karena hanya member yang boleh menambah data cart
+        if($user->user!=2)
+            abort(403);
+        // mengambil data dari table user
         $temp = Cart::where('user_id','=',$user->id)->where('pizza_id','=',$id)->get();
+        // mengecek, apakah temp sudah ada atau belum
+        //jika belum, akan dibuat data cart yang baru
         if($temp->isEmpty()){
             $request->validate([
                 'quantity' => 'required',
@@ -50,6 +67,7 @@ class CartController extends Controller
                 'quantity' => $request->quantity,
             ]);
         }
+        //jika sudah, akan diupdate dari data cart yang sudah ada
         else{
             $request->validate([
                 'quantity' => 'required',
@@ -58,6 +76,7 @@ class CartController extends Controller
                 'quantity' => $temp[0]['quantity']+$request->quantity,
             ]);
         }
+        // berpindah halaman ke halaman cart
         return redirect('/cart');
     }
 
@@ -92,14 +111,23 @@ class CartController extends Controller
      */
     public function update(Request $request, $id)
     {
+        // mengambil data user yang sedang login
+        $user = Auth::user();
+        // mengecek data user, karena hanya member/admin yang boleh berkunjung
+        if($user==null)
+            return view('auth.login');
+        // mengecek data user, karena hanya member yang boleh menambah data cart
+        if($user->user!=2)
+            abort(403);
+        // mengecek quantity pizza yang dimasukan member
         $request->validate([
             'quantity' => 'required',
         ]);
-        $user = Auth::user();
+        // memasukan data ke dalam database
         Cart::where('user_id','=',$user->id)->where('pizza_id','=',$id)->update([
             'quantity' => $request->quantity,
         ]);
-
+        // berpindah halaman ke halaman cart
         return redirect('/cart');
     }
 
@@ -111,8 +139,18 @@ class CartController extends Controller
      */
     public function destroy($id)
     {
+        // mengambil data user yang sedang login
         $user = Auth::user();
+        // mengecek data user, karena hanya member/admin yang boleh berkunjung
+        if($user==null)
+            return view('auth.login');
+        // mengecek data user, karena hanya member yang boleh menambah data cart
+        if($user->user!=2)
+            abort(403);
+        // mendelete data pizza dari table cart sesuai id pizza dan sesuai dengan id user
         Cart::where('user_id','=',$user->id)->where('pizza_id','=',$id)->delete();
+        // berpindah halaman ke halaman pizza
         return redirect('/cart');
     }
+    
 }
