+++ app/Http/Controllers/PizzaController.php
@@ -4,10 +4,26 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Pizza;
+use Auth;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\DB;
 
 class PizzaController extends Controller
 {
+    
+    public function search(Request $request)
+    {
+		// menyimpan data pencarian
+		$search = $request->search;
+ 
+    	// mengambil data dari table pizza sesuai pencarian data
+		$pizzas = DB::table('pizzas')
+		->where('name','like',"%".$search."%")
+		->paginate(5);
+ 
+    	// mengirim data pizza ke view index
+		return view('pizza.index',compact('pizzas'));
+    }
     /**
      * Display a listing of the resource.
      *
@@ -15,7 +31,10 @@ class PizzaController extends Controller
      */
     public function index()
     {
-        $pizzas = Pizza::all();
+    	// mengambil data dari table pizza
+        $pizzas = Pizza::paginate(5);
+        
+    	// mengirim data pizza ke view index
         return view('pizza.index', compact('pizzas'));
     }
 
@@ -26,6 +45,9 @@ class PizzaController extends Controller
      */
     public function create()
     {
+    	// mengecek data user, karena hanya admin yang boleh menambah data pizza baru
+        if(Auth::user()->user!=1)
+            abort(403);
         return view('pizza.create');
     }
 
@@ -37,14 +59,21 @@ class PizzaController extends Controller
      */
     public function store(Request $request)
     {
+        // mengecek data user, karena hanya admin yang boleh menambah data pizza baru
+        if(Auth::user()->user!=1)
+            abort(403);
+        // mengecek data baru pizza yang dimasukan admin
         $request->validate([
             'name' => 'required|max:255|min:3',
             'description' => 'required|min:10',
             'price' => 'required|digits_between:5,6',
             'image' => 'required|mimes:jpeg,png,jpg',
         ]);
+        // membuat nama custom image
         $imgName = $request->image->getClientOriginalName() . '-' . time() . '.' . $request->image->extension();
+        // menyimpan image ke folder public->image
         $request->image->move(public_path('image'), $imgName);
+        // memasukan data ke dalam database
         Pizza::create([
             'name' => $request->name,
             'description' => $request->description,
@@ -53,6 +82,7 @@ class PizzaController extends Controller
             'slug' => Str::slug($request->name,'-')
         ]);
 
+        // berpindah halaman ke halaman pizza
         return redirect('/pizza');
     }
 
@@ -64,10 +94,12 @@ class PizzaController extends Controller
      */
     public function show($slug)
     {
+    	// mengambil data dari table pizza
         $pizza = Pizza::where('slug', $slug)->first();
+    	// mengecek apakah data pizza tersedia atau tidak
         if($pizza == null)
             abort(404);
-
+    	// mengirim data pizza ke halaman detail
         return view('pizza.detail', compact('pizza'));
     }
 
@@ -79,7 +111,12 @@ class PizzaController extends Controller
      */
     public function edit($id)
     {
+        // mengecek data user, karena hanya admin yang boleh mengedit data pizza
+        if(Auth::user()->user!=1)
+            abort(403);
+        // mengambil data dari table pizza sesuai id pizza
         $pizza = Pizza::find($id);
+    	// mengirim data pizza ke halaman edit
         return view('pizza.edit', compact('pizza'));
     }
 
@@ -92,20 +129,28 @@ class PizzaController extends Controller
      */
     public function update(Request $request, $id)
     {
+        // mengecek data user, karena hanya admin yang boleh update data pizza
+        if(Auth::user()->user!=1)
+            abort(403);
+        // mengecek data pizza yang diedit admin
         $request->validate([
             'name' => 'required|max:255|min:3',
             'description' => 'required|min:10',
             'price' => 'required|digits_between:5,6',
             'image' => 'required|mimes:jpeg,png,jpg',
         ]);
+        // membuat nama custom image
         $imgName = $request->image->getClientOriginalName() . '-' . time() . '.' . $request->image->extension();
+        // menyimpan image ke folder public->image
         $request->image->move(public_path('image'), $imgName);
+        // memasukan data ke dalam database
         Pizza::find($id)->update([
             'name' => $request->name,
             'description' => $request->description,
             'price' => $request->price,
             'image' => $imgName
         ]);
+        // berpindah halaman ke halaman pizza
         return redirect('/pizza');
     }
 
@@ -117,7 +162,12 @@ class PizzaController extends Controller
      */
     public function destroy($id)
     {
+        // mengecek data user, karena hanya admin yang boleh delete data pizza
+        if(Auth::user()->user!=1)
+            abort(403);
+        // mendelete data dari table pizza sesuai id pizza
         Pizza::find($id)->delete();
+        // berpindah halaman ke halaman pizza
         return redirect('/pizza');
     }
 }
